fix: updated naming conversion of useThemeContext to useThemeScope
