+++ packages/framework/src/hooks/useThemeContext.tsx
@@ -1,7 +1,10 @@
-import React, { createContext, useContext } from "react";
+import { useAtom, useAtomValue } from "jotai";
+import { clone } from "lodash-es";
+import React, { createContext, useCallback, useContext } from "react";
 import { type EnsembleThemeModel } from "../shared";
+import { appAtom, selectedThemeNameAtom, themeAtom } from "../state";
 
-export type CustomTheme = Record<string, unknown> & {
+export type CustomTheme = { [key: string]: unknown } & {
   theme?: EnsembleThemeModel;
   themeName?: string;
   setTheme?: (name: string) => void;
@@ -22,6 +25,32 @@ export const CustomThemeProvider: React.FC<CustomThemeProps> = ({
   );
 };
 
-export const useCustomTheme = (): CustomTheme => {
+export const useThemeScope = (): {
+  theme: EnsembleThemeModel | undefined;
+  themeName: string;
+  setTheme: (name: string) => void;
+} => {
+  const [theme, updateTheme] = useAtom(themeAtom);
+  const [themeName, updateThemeName] = useAtom(selectedThemeNameAtom);
+  const appContext = useAtomValue(appAtom);
+
+  const setTheme = useCallback(
+    (name: string) => {
+      updateThemeName(name);
+
+      let newTheme = appContext.application?.theme;
+      if (appContext.application?.themes) {
+        newTheme = appContext.application.themes[name];
+      }
+
+      updateTheme(clone(newTheme));
+    },
+    [updateTheme, updateThemeName],
+  );
+
+  return { theme, themeName, setTheme };
+};
+
+export const useTheme = (): CustomTheme => {
   return useContext(CustomThemeContext);
 };
