+++ packages/framework/src/state/application.ts
@@ -42,29 +42,3 @@ export const selectedThemeNameAtom = atomWithStorage<string>(
 );
 
 export const envAtom = focusAtom(appAtom, (optic) => optic.prop("env"));
-
-export const useThemeContext = (): {
-  theme: EnsembleThemeModel | undefined;
-  themeName: string;
-  setTheme: (name: string) => void;
-} => {
-  const [theme, updateTheme] = useAtom(themeAtom);
-  const [themeName, updateThemeName] = useAtom(selectedThemeNameAtom);
-  const appContext = useAtomValue(appAtom);
-
-  const setTheme = useCallback(
-    (name: string) => {
-      updateThemeName(name);
-
-      let newTheme = appContext.application?.theme;
-      if (appContext.application?.themes) {
-        newTheme = appContext.application.themes[name];
-      }
-
-      updateTheme(clone(newTheme));
-    },
-    [updateTheme, updateThemeName],
-  );
-
-  return { theme, themeName, setTheme };
-};
